src/beak/remote/docker/requirements.txt
torch
transformers
numpy
biopython
pandas